refactor(credreceive): added actionsheet & strings & minor refactor of container
