+++ src/ui/home/containers/credentialReceive.tsx
@@ -1,8 +1,13 @@
 import React, { useState } from 'react'
 import { connect } from 'react-redux'
-import { CredentialDialogComponent } from '../components/credentialDialog'
-import { ButtonSection } from 'src/ui/structure/buttonSectionBottom'
-import { View } from 'react-native'
+import {
+  Image,
+  ScrollView,
+  StyleSheet,
+  Text,
+  TouchableOpacity,
+  View,
+} from 'react-native'
 import { ThunkDispatch } from '../../../store'
 import { acceptSelectedCredentials } from '../../../actions/sso/credentialOfferRequest'
 import { CredentialMetadataSummary } from '../../../lib/storage/storage'
@@ -14,6 +19,49 @@ import { JSONWebToken } from 'jolocom-lib/js/interactionTokens/JSONWebToken'
 import { withErrorScreen, withLoading } from '../../../actions/modifiers'
 import { navigationActions } from '../../../actions'
 import { routeList } from '../../../routeList'
+import { DocumentCard } from '../../documents/components/documentCard'
+import { JolocomButton, Wrapper } from '../../structure'
+import { Colors, Spacing } from '../../../styles'
+import { centeredText, fontMain } from '../../../styles/typography'
+import { black065, overflowBlack } from '../../../styles/colors'
+import { ActionSheet } from '../../structure/actionSheet'
+import strings from '../../../locales/strings'
+import I18n from 'src/locales/i18n'
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: Colors.lightGreyLighter,
+  },
+  topSection: {
+    paddingVertical: Spacing.MD,
+    alignItems: 'center',
+  },
+  logo: {
+    borderRadius: 35,
+    width: 70,
+    height: 70,
+    margin: 10,
+  },
+  serviceName: {
+    fontFamily: fontMain,
+    fontSize: 28,
+    color: overflowBlack,
+  },
+  description: {
+    fontFamily: fontMain,
+    fontSize: 16,
+    color: black065,
+    marginTop: 10,
+    marginBottom: 4,
+    marginHorizontal: '10%',
+    ...centeredText,
+  },
+  documentWrapper: {
+    alignItems: 'center',
+    padding: 10,
+  },
+})
 
 interface CredentialNavigationParams {
   credentialOfferRequest: JSONWebToken<CredentialOfferRequest>
@@ -40,6 +88,7 @@ export const CredentialsReceiveContainer = (props: Props) => {
       },
     },
   } = navigation
+  const publicProfile = requesterSummary && requesterSummary.publicProfile
 
   const handleConfirm = () => {
     if (selected.length) {
@@ -68,25 +117,51 @@ export const CredentialsReceiveContainer = (props: Props) => {
   }
 
   return (
-    <View style={{ flex: 1 }}>
-      <View style={{ flex: 0.9 }}>
-        <CredentialDialogComponent
-          requesterSummary={requesterSummary}
-          offerMetadata={offerMetadata}
-          onPressDocument={onPressDocument}
-          selectedCredentials={selected}
+    <Wrapper style={{ backgroundColor: Colors.iBackgroundWhite }}>
+      <ScrollView style={{ width: '100%' }}>
+        {publicProfile && (
+          <View style={styles.topSection}>
+            <Image style={styles.logo} source={{ uri: publicProfile.image }} />
+            <Text style={styles.serviceName}>{publicProfile.name}</Text>
+            <Text style={styles.description}>
+              {I18n.t(
+                strings.CHOOSE_ONE_OR_MORE_DOCUMENTS_PROVIDED_BY_THIS_SERVICE_AND_WE_WILL_GENERATE_THEM_FOR_YOU,
+              )}
+            </Text>
+          </View>
+        )}
+        {offerMetadata.map(summary => {
+          const { type, renderInfo } = summary
+          const isSelected = selected.includes(summary)
+          return (
+            <TouchableOpacity
+              onPress={() => onPressDocument(summary)}
+              activeOpacity={1}
+              style={styles.documentWrapper}
+            >
+              <DocumentCard
+                selected={isSelected}
+                credentialType={type}
+                renderInfo={renderInfo}
+              />
+            </TouchableOpacity>
+          )
+        })}
+      </ScrollView>
+      <ActionSheet showSlide={true}>
+        <JolocomButton
+          disabled={selected.length === 0}
+          onPress={handleConfirm}
+          text={I18n.t(strings.RECEIVE)}
         />
-      </View>
-      <View style={{ flex: 0.1 }}>
-        <ButtonSection
-          confirmText={'Accept'}
-          denyText={'Deny'}
-          handleConfirm={handleConfirm}
-          handleDeny={goBack}
-          disabled={false}
+        <JolocomButton
+          containerStyle={{ marginTop: 10 }}
+          onPress={goBack}
+          text={I18n.t(strings.CANCEL)}
+          transparent
         />
-      </View>
-    </View>
+      </ActionSheet>
+    </Wrapper>
   )
 }
